backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic[email]==2.10.6
python-dotenv==1.0.1
supabase==2.6.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0